requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile requirements.in
#
-e file:///Users/bhuwanbhatt/Documents/open-source/jarvis
    # via -r requirements.in
annotated-types==0.6.0
    # via pydantic
anyio==4.3.0
    # via
    #   httpx
    #   openai
certifi==2024.2.2
    # via
    #   httpcore
    #   httpx
distro==1.9.0
    # via openai
exceptiongroup==1.2.0
    # via anyio
h11==0.14.0
    # via httpcore
httpcore==1.0.5
    # via httpx
httpx==0.27.0
    # via openai
idna==3.6
    # via
    #   anyio
    #   httpx
markdown-it-py==3.0.0
    # via rich
mdurl==0.1.2
    # via markdown-it-py
openai==1.16.0
    # via -r requirements.in
pydantic==2.6.4
    # via openai
pydantic-core==2.16.3
    # via pydantic
pygments==2.17.2
    # via rich
python-dotenv==1.0.1
    # via -r requirements.in
rich==13.7.1
    # via -r requirements.in
sniffio==1.3.1
    # via
    #   anyio
    #   httpx
    #   openai
tqdm==4.66.2
    # via openai
typing-extensions==4.10.0
    # via
    #   anyio
    #   openai
    #   pydantic
    #   pydantic-core
